# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Levy Collection Practice Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # MANAGING AGENTS
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: matterId
          in: query
          schema:
            type: string
        - name: assigneeId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: myTasks
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/complete:
    patch:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteTaskRequest"
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  # REPORTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - isActive
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
      required:
        - email
        - name
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        password:
          type: ["string", "null"]
...
    Task:
      type: object
      properties:
        id:
          type: string
        matterId:
          type: string
        matterReference:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
        assigneeId:
          type: string
        assigneeName:
          type: ["string", "null"]
        isAutoGen:
          type: boolean
        completedAt:
          type: ["string", "null"]
        completionNote:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - matterId
        - title
        - status
        - priority
        - assigneeId
        - isAutoGen
        - createdAt
    CreateTaskRequest:
      type: object
      properties:
        matterId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
        assigneeId:
          type: string
      required:
        - matterId
        - title
        - priority
        - assigneeId
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
        assigneeId:
          type: string
        status:
          type: string
    CompleteTaskRequest:
      type: object
      properties:
        note:
          type: ["string", "null"]
...